.gitignore: Add `/fish/fish_config/fish_variables`
This is where the fish "universal" variables are stored. These are
variables that are applied to all open fish shells.

In theory, something that we'd want to store or "save" could be in here
but at the moment nothing is and principally secrets could be in here
as the variables are often "adjacent" to environment variables
(the universal variables can be read as environment variables).

For variables that we do want to store (e.g. `$EDITOR`), we should set
them in `fish_config`.