Hammerspoon[window-drag]: Only raise and focus window once mouse moved
Previously, when pressing certain keyboard shortcuts in a different
window we would raise and focus the window under the cursor which would
be some irrelevant background window to the one we were typing in.

We now do something smarter where we only focus and raise the window
once the cursor is moved -- which would indicate and actual window move
or resize.